Allow Restarting Triggerer and Scheduler separately
Since during development, some of the changes in files don't get used due to Python cache and the cache used in the TriggererJob, we will restart those containers as this is a very quick operator.

Restarting Webserver take a long time but is not needed similar to other components.

To restart all components, we can use `make restart-all`