Fix EMR EKS example DAG (#259)
* Fail the shell script if any of the commands fail

Currently, the shell script is silently ignoring failure of
commands. As a result, the Airflow task calling this shell script
is marked as 'success' which should not happen because the DAG
proceeds to execute subsequent tasks resulting in late failures.
With this change, we want to report error in any command upfront
so that the task is marked 'failed' and the DAG does not continue
executing downstream tasks.

* Add separate Dockerfile for EMR-EKS example DAG image

* Update astro-runtime image tag to latest version

* Pass needed environment variables to shell script

The shell script to create and delete EKS cluster along with other
resources needs a set of environment variables which were not
being passed from the respective Airflow operators. These are being
passed now.
The change also orders the variables alphabetically.

Additionally, the 'create_emr_virtual_cluster' task generates
the cluster ID which needs to be passed to subserquent tasks. For
this purpose, we push the ID to xcom and the subsequent tasks pull
it from the XCOM database.

* Install required packages for EMR EKS DAG image