Reraise exceptions caught in except block (#338)
Currently, we're catching few exceptions in our tasks as part of our example DAGs.
Such exceptions caught do log the exceptions but return None. As a result,
the task is getting marked as success/green and the downstream tasks fail
due to missing resources of this previous task outcome.
Fix such exceptions caught to rightly reflect the status of the tasks.