Fix Redshift DAGs to catch appropriate exceptions (#348)
* Cluster delete and snapshot delete tasks keep on waiting while
the status is in 'deleting' state. Since, the status get is part
of a while loop, the last iteration when they get deleted raises
an exception that the corresponding resource is not found and it
marks the task as failed. We are fixing this by catching the relevant
status code and other exceptions are re-raised.

Additionally, it is observed quite often that the DAG tasks fail
without any logs. @ephraimbuddy suggested that this could be due
to DAG processing timeouts occuring due to time spent in importing
heavy libraries. So taking reference of
https://airflow.apache.org/docs/apache-airflow/stable/best-practices.html#top-level-python-code
we're delaying the import of 'boto' to the task execution stage and
avoiding to it import it at top module level.

We're also also renaming the operator's python reference variable
names to be consistent with the 'task_id'.

* Add logs to indicate sleep